requirements.txt
flask
whisper
gtts
pydub
moviepy
opencv-python
ffmpeg-python
pillow
